# cmd/api/docs/swagger.yaml
basePath: /
definitions:
...
  model.Certification:
    properties:
      certification_id:
        type: integer
      date_issued:
        type: string
      issuer:
        type: string
      name:
        type: string
      profile_id:
        type: integer
    type: object
  model.Contact:
    properties:
      contact_id:
        type: integer
      email:
        type: string
      phone_number:
        type: string
      profile_id:
        type: integer
    type: object
  model.Education:
    properties:
      degree:
        type: string
      education_id:
        type: string
      end_date:
        type: string
      field:
        type: string
      institution:
        type: string
      profile_id:
        type: integer
      start_date:
        type: string
    type: object
  model.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API para gerenciamento de perfis e currículos com IA.
  title: Resume GenAI API
  version: "1.0"
paths:
  /certifications/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Certification'
            type: array
      summary: Lista todas as certificações
      tags:
      - certifications
    post:
      consumes:
      - application/json
      parameters:
      - description: Certificação a ser criada
        in: body
        name: certification
        required: true
        schema:
          $ref: '#/definitions/model.Certification'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Certification'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Cria uma nova certificação
      tags:
      - certifications
  /certifications/{id}:
    delete:
      parameters:
      - description: ID da certificação
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Remove uma certificação
      tags:
      - certifications
    get:
      parameters:
      - description: ID da certificação
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Certification'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Busca certificação por ID
      tags:
      - certifications
    put:
      consumes:
      - application/json
      parameters:
      - description: ID da certificação
        in: path
        name: id
        required: true
        type: string
      - description: Certificação atualizada
        in: body
        name: certification
        required: true
        schema:
          $ref: '#/definitions/model.Certification'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Certification'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Atualiza uma certificação
      tags:
      - certifications
  /contacts/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Contact'
            type: array
      summary: Lista todos os contatos
      tags:
      - contacts
    post:
      consumes:
      - application/json
      parameters:
      - description: Contato a ser criado
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/model.Contact'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Contact'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Cria um novo contato
      tags:
      - contacts
  /contacts/{id}:
    delete:
      parameters:
      - description: ID do contato
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Remove um contato
      tags:
      - contacts
    get:
      parameters:
      - description: ID do contato
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Contact'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Busca contato por ID
      tags:
      - contacts
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do contato
        in: path
        name: id
        required: true
        type: integer
      - description: Contato atualizado
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/model.Contact'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Contact'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Atualiza um contato
      tags:
      - contacts
  /educations/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Education'
            type: array
      summary: Lista todas as formações
      tags:
      - educations
    post:
      consumes:
      - application/json
      parameters:
      - description: Formação a ser criada
        in: body
        name: education
        required: true
        schema:
          $ref: '#/definitions/model.Education'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Education'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Cria uma nova formação
      tags:
      - educations
  /educations/{id}:
    delete:
      parameters:
      - description: ID da formação
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Remove uma formação
      tags:
      - educations
    get:
      parameters:
      - description: ID da formação
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Education'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Busca formação por ID
      tags:
      - educations
    put:
      consumes:
      - application/json
      parameters:
      - description: ID da formação
        in: path
        name: id
        required: true
        type: string
      - description: Formação atualizada
        in: body
        name: education
        required: true
        schema:
          $ref: '#/definitions/model.Education'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Education'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Atualiza uma formação
      tags:
      - educations
...
swagger: "2.0"